Make cors work better.
The way cors is set up does not play well with certain applications for swift.
If you are running a CDN on top of swift and you have the
Access-Control-Allow-Origin cors header set to * then you probably want the *
to be cached on the the CDN, not the Origin that happened to result in an
origin request.

Also- if were misguided enough to allow cors headers to be saved directly onto
objects then this allows them to supersede the headers coming from the
container.

Change-Id: I9b0219407e77c77a9bb1133cbcb179a4c681c4a8