xLO bug with auth tokens expiring during download.
With the xLO move to middleware, every segment GET was subjected to auth. While
this provided better security than the previous implementation, during long
downloads the client can get incomplete results if their auth token expires.

With this change, subsegment requests will only be authed if it is the first
request to the given container. If all the segments are in the same container,
as is always the case with DLOs, then only the first segment request will be
authed.

There is also a new "swift.cache_authentication" environment var added that an
auth middleware can respond to to cache the authentication of a request in the
wsgi environment. Basically, the auth middleware will only have to hit the
authentication service once for the request but will still honor the
authorization (container permission ACL, etc) for each of the subrequests. If
the auth middleware honors this, then SLOs with segments in multiple containers
can have a token that won't expire while being able to lock users out of
reading from containers they don't have permission to. If the auth middleware
does not honor this then users with segments in multiple containers have a
chance of getting a 401 on a segment request if the token expires before
pulling segments from all the needed containers.

Change-Id: I7827dd2d9dfbb3c6424773fb2891355d47e372ba