Add container wide Cache-Control for static web containers
If you are using static web / public containers to use swift as an origin
server it would be nice to be able to set Cache-Control at the container
level instead of having to set it per object with custom object headers. This
change will only set Cache-Control if it not already set by custome object
metadata.

Change-Id: I3292e301e9bd92e925bc9e19ac52d66a419beddf